requirements.txt
streamlit
plotly
pandas
pypdf2
python-docx
pillow
nbformat
zipfile
